Move highcharts package to cmp package, rename 'charts'
