+++ cmp/chart/index.js
